docs(852): add composed -> the sun
